refactoring: move tools helper into provider-utils
