+++ .changeset/tough-suns-learn.md
@@ -1,0 +1,5 @@
+---
+'@ai-sdk/provider-utils': patch
+---
+
+refactoring: move tools helper into provider-utils
