+++ packages/ai/core/prompt/convert-to-language-model-prompt.test.ts
@@ -1,4 +1,3 @@
-import { tool } from '../tool/tool';
 import {
   convertToLanguageModelMessage,
   convertToLanguageModelPrompt,
