+++ packages/ai/core/prompt/prepare-tools-and-tool-choice.test.ts
@@ -1,8 +1,7 @@
 import { z } from 'zod';
 import { ToolSet } from '../generate-text/tool-set';
-import { tool } from '../tool/tool';
 import { prepareToolsAndToolChoice } from './prepare-tools-and-tool-choice';
-import { Tool } from '@ai-sdk/provider-utils';
+import { Tool, tool } from '@ai-sdk/provider-utils';
 
 const mockTools: ToolSet = {
   tool1: tool({
