+++ packages/ai/core/tool/index.ts
@@ -7,13 +7,13 @@ export type {
 } from './mcp/json-rpc-message';
 export { createMCPClient as experimental_createMCPClient } from './mcp/mcp-client';
 export type { MCPTransport } from './mcp/mcp-transport';
-export { tool } from './tool';
 
-// re-export types from provider-utils
-export type {
-  Tool,
-  ToolCallOptions,
-  ToolExecuteFunction,
-  InferToolInput,
-  InferToolOutput,
+// re-exports from provider-utils
+export {
+  tool,
+  type Tool,
+  type ToolCallOptions,
+  type ToolExecuteFunction,
+  type InferToolInput,
+  type InferToolOutput,
 } from '@ai-sdk/provider-utils';
