+++ packages/provider-utils/src/provider-defined-client-tool-factory.ts
@@ -1,4 +1,4 @@
-import { Tool, ToolExecuteFunction } from './types/tool';
+import { tool, Tool, ToolExecuteFunction } from './types/tool';
 import { FlexibleSchema } from './schema';
 
 export type ProviderDefinedClientToolFactory<INPUT, ARGS extends object> = <
@@ -20,11 +20,12 @@ export function createProviderDefinedClientToolFactory<
   id,
   inputSchema,
 }: {
-  id: string;
+  id: `${string}.${string}`;
   inputSchema: FlexibleSchema<INPUT>;
 }): ProviderDefinedClientToolFactory<INPUT, ARGS> {
   return <OUTPUT>({
     execute,
+    outputSchema,
     toModelOutput,
     onInputStart,
     onInputDelta,
@@ -32,20 +33,22 @@ export function createProviderDefinedClientToolFactory<
     ...args
   }: ARGS & {
     execute?: ToolExecuteFunction<INPUT, OUTPUT>;
+    outputSchema?: FlexibleSchema<OUTPUT>;
     toModelOutput?: Tool<INPUT, OUTPUT>['toModelOutput'];
     onInputStart?: Tool<INPUT, OUTPUT>['onInputStart'];
     onInputDelta?: Tool<INPUT, OUTPUT>['onInputDelta'];
     onInputAvailable?: Tool<INPUT, OUTPUT>['onInputAvailable'];
   }): Tool<INPUT, OUTPUT> =>
-    ({
+    tool({
       type: 'provider-defined-client',
       id,
       args,
       inputSchema,
+      outputSchema,
       execute,
       toModelOutput,
       onInputStart,
       onInputDelta,
       onInputAvailable,
-    }) as Tool<INPUT, OUTPUT>;
+    });
 }
