+++ packages/provider-utils/src/types/index.ts
@@ -14,12 +14,13 @@ export type { DataContent } from './data-content';
 export type { ModelMessage } from './model-message';
 export type { ProviderOptions } from './provider-options';
 export type { SystemModelMessage } from './system-model-message';
-export type {
-  Tool,
-  ToolCallOptions,
-  ToolExecuteFunction,
-  InferToolInput,
-  InferToolOutput,
+export {
+  tool,
+  type Tool,
+  type ToolCallOptions,
+  type ToolExecuteFunction,
+  type InferToolInput,
+  type InferToolOutput,
 } from './tool';
 export type { ToolCall } from './tool-call';
 export type { ToolContent, ToolModelMessage } from './tool-model-message';
