+++ packages/provider-utils/src/types/tool.ts
@@ -167,3 +167,17 @@ export type InferToolInput<TOOL extends Tool> =
  */
 export type InferToolOutput<TOOL extends Tool> =
   TOOL extends Tool<any, infer OUTPUT> ? OUTPUT : never;
+
+/**
+Helper function for inferring the execute args of a tool.
+ */
+// Note: overload order is important for auto-completion
+export function tool<INPUT, OUTPUT>(
+  tool: Tool<INPUT, OUTPUT>,
+): Tool<INPUT, OUTPUT>;
+export function tool<INPUT>(tool: Tool<INPUT, never>): Tool<INPUT, never>;
+export function tool<OUTPUT>(tool: Tool<never, OUTPUT>): Tool<never, OUTPUT>;
+export function tool(tool: Tool<never, never>): Tool<never, never>;
+export function tool(tool: any): any {
+  return tool;
+}
